python: Fix issues with S3 upload/download
Specify Content-Type otherwise it gets assigned 'binary-stream' and
the web will return download link to index only.

Check if 'czechia_table' exists in the bucket. This is likely problem
only during the first run. If it does not, we create the table from new data.

Do not display or process index column (first column in COVID19 data).
Not very useful information to display and pandas creates 'Unnamed'
column for it.